LLVM10 and rework llvm deps

+++ deps/llvm/build.py
@@ -7,74 +7,106 @@ import shutil
 import sys
 import tarfile
 import itertools
+import hashlib
+import platform
 from pathlib import Path
 try:
     from urllib.request import urlopen
 except ImportError:
     raise Exception("Must be using Python 3")
 
-VERSION = "8.0.1"
+VERSION = "10.0.1"
 #SOURCE_URL = "http://llvm.org/releases/" + VERSION + "/llvm-" + \
 #    VERSION + ".src.tar.xz"
 
 SOURCE_URL = "https://github.com/llvm/llvm-project/releases/download/llvmorg-{}/llvm-{}.src.tar.xz" \
         .format(VERSION, VERSION)
 
+SOURCE_SHA256 = "c5d8e30b57cbded7128d78e5e8dad811bff97a8d471896812f57fa99ee82cdf3"
+
 TARGET_DIR = Path(os.path.dirname(__file__))
 
 
-def get_libraries(target_arch, prefix="", ext=""):
+def get_libraries(target_arch, target_platform, prefix="", ext=""):
     """
     Get the list of libraries required from LLVM for QBDI.
 
     It is only libraries names, provide a prefix and an extension if required.
     """
-    # FIXME: Do we need Demangle for every platform?
-    LIBRAIRIES = ["LLVMSelectionDAG", "LLVMAsmPrinter",
-                  "LLVMBinaryFormat", "LLVMCodeGen",
-                  "LLVMScalarOpts", "LLVMProfileData",
-                  "LLVMInstCombine", "LLVMTransformUtils",
-                  "LLVMAnalysis", "LLVMTarget", "LLVMObject",
-                  "LLVMMCParser", "LLVMBitReader",
-                  "LLVMMCDisassembler", "LLVMMC",
-                  "LLVMCore", "LLVMSupport", "LLVMDemangle"]
+    LIBRAIRIES = [
+        # needed by QBDI
+        "LLVMBinaryFormat",
+        "LLVMMCDisassembler",
+        "LLVMMCParser",
+        "LLVMMC",
+        "LLVMSupport",
+        # needed by QBDITest
+        "LLVMObject",
+        "LLVMTextAPI",
+        "LLVMCore",
+        "LLVMBitReader",
+        "LLVMBitstreamReader",
+        "LLVMRemarks",
+        ]
     # Add target specific libraries
-    for target_lib in ("Utils", "Info", "Disassembler", "Desc", "CodeGen",
-                       "AsmPrinter", "AsmParser"):
+    if target_platform in ['iOS', 'macOS']:
+        LIBRAIRIES.append("LLVMDemangle")
+
+    for target_lib in ("Utils", "Info", "Disassembler", "Desc", "AsmParser"):
         LIBRAIRIES.append("LLVM" + target_arch + target_lib)
+
     return [prefix + l + ext for l in LIBRAIRIES]
 
+def verify_hash(filename, expected_hash):
 
-def get_llvm(url, dest):
+    m = hashlib.sha256()
+    with filename.open('rb') as f:
+        m.update(f.read())
+
+    return m.hexdigest() == expected_hash
+
+def get_llvm(url, source_hash, dest):
     """Download and extract llvm."""
     # Force removal even if download should remove it too.
     base_name = dest / os.path.basename(url)
-    if base_name.exists():
-        os.remove(base_name)
+    should_download = True
 
-    # FIXME: We should use a specific target path
-    print("Downloading llvm ...")
-    f = urlopen(url)
-    with base_name.open("wb") as tmp_f:
-        tmp_f.write(f.read())
+    if base_name.exists():
+        print("Verify existing file hash ...")
+        if verify_hash(base_name, source_hash):
+            should_download = False
+        else:
+            print("hash mismatch, redownload ...")
+            os.remove(base_name)
+
+    if should_download:
+        # FIXME: We should use a specific target path
+        print("Downloading llvm ...")
+        f = urlopen(url)
+        with base_name.open("wb") as tmp_f:
+            tmp_f.write(f.read())
+
+        print("Verify hash ...")
+        if not verify_hash(base_name, source_hash):
+            RuntimeError("Hash verifucation failed")
 
     print("Extract llvm ...")
     with tarfile.open(str(base_name), 'r|xz') as tar:
         tar.extractall(str(dest))
 
 
-def build_llvm(llvm_dir, build_dir, arch, platform, arch_opt=None):
+def build_llvm(llvm_dir, build_dir, arch, target_plateform, arch_opt=None):
 
     if not build_dir.exists():
         build_dir.mkdir()
 
     # set platform specific arguments.
-    if platform == "win" and arch_opt == "i386":
+    if target_plateform == "win" and arch_opt == "i386":
         cmake_specific_option = ["-G", "Ninja",
                                  "-DLLVM_BUILD_32_BITS=On"]
-    elif platform == "win":
+    elif target_plateform == "win":
         cmake_specific_option = ["-G", "Ninja"]
-    elif platform == "iOS":
+    elif target_plateform == "iOS":
         cc = subprocess.check_output(["xcrun", "--sdk", "iphoneos", "-f", "clang"])
         cxx = subprocess.check_output(["xcrun", "--sdk", "iphoneos", "-f", "clang++"])
         sdk = subprocess.check_output(["xcrun", "--sdk", "iphoneos", "--show-sdk-path"])
@@ -87,7 +119,7 @@ def build_llvm(llvm_dir, build_dir, arch, platform, arch_opt=None):
                                  "-DCMAKE_CXX_FLAGS=-arch " + arch_opt + " -isysroot " + sdk + " -fvisibility=hidden",
                                  "-DCMAKE_C_COMPILER=" + cc,
                                  "-DCMAKE_CXX_COMPILER=" + cxx]
-    elif platform == "linux" and arch == "ARM":
+    elif target_plateform == "linux" and arch == "ARM":
         ARM_C_INCLUDE="/usr/arm-linux-gnueabi/include/"
         ARM_CXX_INCLUDE="/usr/arm-linux-gnueabi/include/c++/6/"
         cmake_specific_option = [
@@ -98,7 +130,7 @@ def build_llvm(llvm_dir, build_dir, arch, platform, arch_opt=None):
             "-DCMAKE_C_FLAGS=-fvisibility=hidden -march=" + arch_opt + " -I" + ARM_C_INCLUDE,
             "-DCMAKE_CXX_FLAGS=-fvisibility=hidden -march=" + arch_opt + " -I" + ARM_C_INCLUDE + " -I" + ARM_CXX_INCLUDE,
         ]
-    elif platform == "android":
+    elif target_plateform == "android":
         defaul_ndk_path = str(Path.home() / "android-ndk-r20")
         ndk_path = os.getenv("NDK_PATH", defaul_ndk_path)
 
@@ -130,12 +162,12 @@ def build_llvm(llvm_dir, build_dir, arch, platform, arch_opt=None):
                     '-DANDROID_ABI=x86_64',
                     ]
 
-    elif platform == "macOS" and arch_opt == "i386":
+    elif target_plateform == "macOS" and arch_opt == "i386":
         cmake_specific_option = ['-DLLVM_BUILD_32_BITS=On',
                                  '-DLLVM_DEFAULT_TARGET_TRIPLE=i386-apple-darwin17.7.0',
                                  '-DCMAKE_C_FLAGS=-fvisibility=hidden',
                                  '-DCMAKE_CXX_FLAGS=-fvisibility=hidden']
-    elif platform == "linux" and arch_opt == "i386":
+    elif target_plateform == "linux" and arch_opt == "i386":
         cmake_specific_option = ['-DLLVM_BUILD_32_BITS=On',
                                  '-DLLVM_DEFAULT_TARGET_TRIPLE=i386-pc-linux',
                                  '-DCMAKE_C_FLAGS=-fvisibility=hidden',
@@ -151,23 +183,28 @@ def build_llvm(llvm_dir, build_dir, arch, platform, arch_opt=None):
                            "-DLLVM_BUILD_TOOLS=Off",
                            "-DLLVM_BUILD_UTILS=Off",
                            "-DLLVM_BUILD_TESTS=Off",
-                           "-DLLVM_BUILD_EXAMPLES=Off"
-                           "-DLLVM_INCLUDE_TOOLS=Off"
-                           "-DLLVM_INCLUDE_UTILS=Off"
-                           "-DLLVM_INCLUDE_TESTS=Off"
+                           "-DLLVM_BUILD_EXAMPLES=Off",
+                           "-DLLVM_BUILD_BENCHMARKS=Off",
+                           "-DLLVM_INCLUDE_TOOLS=Off",
+                           "-DLLVM_INCLUDE_UTILS=Off",
+                           "-DLLVM_INCLUDE_TESTS=Off",
                            "-DLLVM_INCLUDE_EXAMPLES=Off",
+                           "-DLLVM_INCLUDE_BENCHMARKS=Off",
                            "-DLLVM_ENABLE_TERMINFO=Off",
-                           "-DLLVM_ENABLE_BINDINGS=Off"] +
-                          cmake_specific_option,
+                           "-DLLVM_ENABLE_BINDINGS=Off",
+                           "-DLLVM_ENABLE_Z3_SOLVER=Off",
+                           "-DLLVM_APPEND_VC_REV=Off",
+                           "-DLLVM_ENABLE_RTTI=Off",
+                           ] + cmake_specific_option,
                           cwd=str(build_dir))
 
     # Compile llvm libraries
-    if platform == "win":
-        subprocess.check_call(["ninja"] + get_libraries(arch, ext=".lib"),
+    if target_plateform == "win":
+        subprocess.check_call(["ninja"] + get_libraries(arch, target_plateform, ext=".lib"),
                               cwd=str(build_dir))
     else:
         # FIXME: Not always 4, could be 8 on my computer and 2 on travis)
-        subprocess.check_call(["make", "-j4"] + get_libraries(arch), cwd=str(build_dir))
+        subprocess.check_call(["make", "-j4"] + get_libraries(arch, target_plateform), cwd=str(build_dir))
 
 
 def extract_file(source, dest, exts):
@@ -180,8 +217,7 @@ def extract_file(source, dest, exts):
         shutil.copy(str(header), str(new_header))
 
 
-def install_header_and_lib(llvm_source, build_dir, dest, arch):
-    import platform
+def install_header_and_lib(llvm_source, build_dir, dest, arch, target_plateform):
     shutil.rmtree(str(dest / "lib"), ignore_errors=True)
     shutil.rmtree(str(dest / "include"), ignore_errors=True)
 
@@ -199,7 +235,7 @@ def install_header_and_lib(llvm_source, build_dir, dest, arch):
     lib_path = os.path.join(str(build_dir), "lib")
     PREFIX = "lib" if platform.system() != "Windows" else ""
     SUFFIX = ".lib" if platform.system() == "Windows" else ".a"
-    for l in get_libraries(arch, prefix=PREFIX, ext=SUFFIX):
+    for l in get_libraries(arch, target_plateform, prefix=PREFIX, ext=SUFFIX):
         shutil.copy(os.path.join(lib_path, l), str(dest / "lib"))
 
     if platform.system() == "Darwin":
@@ -229,16 +265,16 @@ if __name__ == "__main__":
     target = sys.argv[2]
     assert(target in ("android-ARM", "android-X86", "android-X86_64", "iOS-ARM", "linux-ARM", "linux-X86_64", "linux-X86", "macOS-X86_64", "macOS-X86", "win-X86_64", "win-X86"))
 
-    platform = target.split("-")[0]
+    target_plateform = target.split("-")[0]
     arch = target.split("-")[1]
 
     arch_opt = None
     if arch == "ARM":
-        if platform == "linux":
+        if target_plateform == "linux":
             arch_opt = "armv6"
-        elif platform == "iOS":
+        elif target_plateform == "iOS":
             arch_opt = "armv7"
-        elif platform == "android":
+        elif target_plateform == "android":
             arch_opt = "armv7-a"
         else:
             raise RuntimeError("Unknown platform")
@@ -254,17 +290,17 @@ if __name__ == "__main__":
     extracted_llvm_dir = TARGET_DIR / target / ("llvm-" + VERSION + ".src")
 
     if sys.argv[1] == "prepare":
-        get_llvm(SOURCE_URL, TARGET_DIR / target)
+        get_llvm(SOURCE_URL, SOURCE_SHA256, TARGET_DIR / target)
 
     elif sys.argv[1] == "build":
         build_llvm(extracted_llvm_dir,
-                   build_dir, arch, platform, arch_opt)
+                   build_dir, arch, target_plateform, arch_opt)
 
     elif sys.argv[1] == "package":
         install_header_and_lib(extracted_llvm_dir,
                                build_dir,
                                TARGET_DIR / target,
-                               arch)
+                               arch, target_plateform)
 
     elif sys.argv[1] == "clean":
         clean_llvm_build(extracted_llvm_dir,
